// src/app/views/prj-planning/model/project.ts
export class Project {
   id?: string;
   name?: string ;
   cost?:  number ;
   office?:  string ;
   projectDirector?: string ;
   startDate?: Date ;
   finishDate?: Date ;
   initiatorOffice?: string[] ;
   implementingAgency?:  string[] ;
   planningCommission?:  string ;
   objective?:  string ;
   createDate?: Date;
   updateDate?: Date;
   projectCode?: string;
   officeId?: string;
   approveDate?: Date;
   activities?: string;
}